data_producer/requirements.txt
Faker==37.4.0
kafka-python==2.2.15
